+++ backend/backend/settings/test.py
@@ -11,6 +11,8 @@ ORG_NAME = "OrgTestSuite"
 DEFAULT_DOMAIN = "http://testserver"
 
 ASSET_BUFFER_DIR = tempfile.mkdtemp()  # overridden in individual tests
+SUBTUPLE_DIR = os.path.join(MEDIA_ROOT, "subtuple")
+SUBTUPLE_TMP_DIR = os.path.join(SUBTUPLE_DIR, "tmp")
 
 ORCHESTRATOR_HOST = "orchestrator"
 ORCHESTRATOR_PORT = 9000