+++ CHANGELOG.md
@@ -7,8 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Fixed
-
 - Memory leak in MinIO client
+
+### Changed
+- When executing compute tasks, store Algos and Metrics in temporary folders instead of the Asset Buffer
 
 ## [0.6.0] 2021-12-01